requirements.txt
openai
python-dotenv
pyautogen